working 32bit build on gcc13

+++ base/applications/atactl/atactl.cpp
@@ -1017,8 +1017,8 @@ ata_adapter_info(
         } else {
             if(AdapterInfo->AdapterInterfaceType == PCIBus) {
                 slotData.u.AsULONG = AdapterInfo->slotNumber;
-                printf("  PCI Bus/Dev/Func:   %lu/%lu/%lu%s\n",
-                    AdapterInfo->SystemIoBusNumber, slotData.u.bits.DeviceNumber, slotData.u.bits.FunctionNumber,
+                printf("  PCI Bus/Dev/Func:   %u/%u/%u%s\n",
+                    (unsigned int)AdapterInfo->SystemIoBusNumber, (unsigned int)slotData.u.bits.DeviceNumber, (unsigned int)slotData.u.bits.FunctionNumber,
                     AdapterInfo->AdapterInterfaceType == AdapterInfo->OrigAdapterInterfaceType ? "" : " (ISA-Bridged)");
                 printf("  VendorId/DevId/Rev: %#04x/%#04x/%#02x\n",
                     (USHORT)(AdapterInfo->DevID >> 16),
